Updated to version 0.9.10.  (minor update, mostly bugfixes and code-cleanup)
 - fixed Enforcer-Hits if no project-file has been specified and
   external URIs have been referenced (reported by Jens T. Berger 
   Thielemann)
 - fixed bug that caused trashed relative URIs, if a <BASE>-tag 
   occured (reported by Andreas Gaszner)
 - fixed bug that allowed attributes of type URI being parsed
   more then once if they've been declared using <$define>,
   usually resulting into wrong paths for existence-check
   (reported by Andreas Gaszner)
 - fixed bug that caused the relative destination-dir to be omitted
   for the output-filename, if a full input-filename has been passed
   (reported by Walter Doerwald)
 - fixed a minor bug in grafflwerk/hScMsg.rexx, which could lead to
   a wrong return-code if ScMsg had to be invoked from the script
 - stopped supporting plain 68000 CPUs; now at least a 68020 is
   required for hsc and hscdepp
 - added CLI option DEFINE
 - added CLI-option MSGANSI
 - added CLI-option PREFSFILE
 - added CLI-option QUOTESTYLE
 - improved id-checking that it now also works if current document
   is referenced indirectly via it's own filename.
 - improved handling of input-files (speed increase for files >100k)
 - improved code-sharing for hscdepp (smaller executable)
 - improved help; it now also shows argument-type and -flags
 - cleaned-up "Makefile.dodl" so that it only contains the
   dependencies really required.
 - compiled with gcc 2.7.0 and libnix 1.1

+++ src/ugly/umemory.c
@@ -21,7 +21,7 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 675 Mass Ave, Cambridge, MA 02139, USA.
  *
- * updated: 30-Jul-1996
+ * updated: 13-Sep-1996
  * created: 29-Mar-1994
  *
  */
@@ -356,7 +356,7 @@ static BOOL ugly_walldamaged(UGLYMEM * umem)
 
             fprintf(stderr, "*** MEMORY WALL DAMAGED!!!\n");
             fprintf(stderr, "*** %s wall, byte#%lu is %s instead of 0x%02x\n",
-                    wall, i, str_ubyte(value), umem->fillchar);
+                    wall, (ULONG) i, str_ubyte(value), umem->fillchar);
             umem_info(umem);
             ugly_memdump(umem->ptr, umem->size);
             fprintf(stderr, "  * lower wall:\n");
